+++ src/index.tsx
@@ -4,13 +4,10 @@ import './tailwind.css';
 import './sass/index.scss';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
-// import PageBuilderApp from './PageBuilderApp';
-// import pageBuilderAnswersAppConfig from './config/pageBuilderAnswersAppConfig';
 
 ReactDOM.render(
   <React.StrictMode>
     <App />
-    {/* <PageBuilderApp config={pageBuilderAnswersAppConfig}/> */}
   </React.StrictMode>,
   document.getElementById('root')
 );
